Data_Workflow/pyproject.toml
[project]
name = "data-workflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "plotly-express>=0.4.1",
    "pyarrow>=22.0.0",
    "nbformat>=4.2.0",
]
